some refactoring that missed the last commit
