+++ src/dotless.Test/dotless.Test.csproj
@@ -126,8 +126,10 @@
     <Compile Include="Specs\ScriptFixture.cs" />
     <Compile Include="TestLogger.cs" />
     <Compile Include="TestStylizer.cs" />
+    <Compile Include="Unit\Bundling\InMemoryVirtualFile.cs" />
+    <Compile Include="Unit\Bundling\InMemoryVirtualPathProvider.cs" />
     <Compile Include="Unit\HttpHandlerFixture.cs" />
-    <Compile Include="Unit\LessTransformerFixture.cs" />
+    <Compile Include="Unit\Bundling\LessTransformerFixture.cs" />
     <Compile Include="Unit\Tokenizer\TokenizerFixture.cs" />
     <Compile Include="Unit\Engine\CacheDecoratorFixture.cs" />
     <Compile Include="Unit\Engine\CommentBug.cs" />
